django_backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
daphne==4.0.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
kombu==5.3.4
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
drf-spectacular==0.26.5
gunicorn==21.2.0
